+++ src/jk_datamatrix/__init__.py
@@ -1,7 +1,10 @@
 ﻿
 
-
-__version__ = "0.2024.6.27"
+__author__ = "Jürgen Knauth"
+__version__ = "0.2025.5.24"
+__email__ = "pubsrc@binary-overflow.de"
+__license__ = "Apache2"
+__copyright__ = "Copyright (c) 2021-2025, Jürgen Knauth"
 
 
 
@@ -15,4 +18,10 @@ from .DataMatrix import DataMatrix
 
 # the loaders. they require an already defined DataMatrix.
 from .DataMatrixJSONLoader import DataMatrixJSONLoader
+
+
+
+# convenience methods
+loadFromJSONFile = DataMatrixJSONLoader.loadFromJSONFile
+loadFromJSONStr = DataMatrixJSONLoader.loadFromJSONStr